fix: Offline handler handles failed requests
