+++ src/utils/featheryClient/integrationClient.ts
@@ -1,15 +1,15 @@
-import * as errors from '../error';
 import { parseError } from '../error';
 import { fieldValues, initFormsPromise, initInfo } from '../init';
 import { encodeGetParams } from '../primitives';
 import { API_URL, STATIC_URL } from '.';
-import { OfflineRequestHandler, untrackUnload } from '../offlineRequestHandler';
+import { OfflineRequestHandler } from '../offlineRequestHandler';
 import {
   AlloyEntities,
   IntegrationActionIds,
   IntegrationActionOptions,
   LoanProCustomerObject
 } from '../internalState';
+import { checkResponseSuccess } from './utils';
 
 export const TYPE_MESSAGES_TO_IGNORE = [
   // e.g. https://sentry.io/organizations/feathery-forms/issues/3571287943/
@@ -17,33 +17,6 @@ export const TYPE_MESSAGES_TO_IGNORE = [
   // e.g. https://sentry.io/organizations/feathery-forms/issues/3529742129/
   'Load failed'
 ];
-
-export async function checkResponseSuccess(response: any) {
-  let payload;
-  switch (response.status) {
-    case 200:
-    case 201:
-    case 202:
-      return;
-    case 400:
-      payload = JSON.stringify(await response.clone().text());
-      console.error(payload.toString());
-      return;
-    case 401:
-      throw new errors.SDKKeyError();
-    case 404:
-      throw new errors.FetchError("Can't find object");
-    case 409:
-      // Note: remove beforeunload listeners if there is a conflict
-      untrackUnload(true);
-      location.reload();
-      return;
-    case 500:
-      throw new errors.FetchError('Internal server error');
-    default:
-      throw new errors.FetchError('Unknown error');
-  }
-}
 
 // THIRD-PARTY INTEGRATIONS
 export default class IntegrationClient {
